+++ traveller-android-app/traveller/src/main/java/com/requestlabs/traveller/MainActivity.java
@@ -68,6 +68,9 @@ public class MainActivity extends ActionBarActivity implements AdapterView.OnIte
 
 
     private String normalizedTaskDescription() {
+        if(mTaskInput == null || mTaskInput.getText() == null){
+            return "";
+        }
         return mTaskInput.getText().toString().trim();
     }